refactor track_refund_status_changed
