+++ app/services/credit_notes/refunds/gocardless_service.rb
@@ -34,7 +34,7 @@ module CreditNotes
         refund.save!
 
         update_credit_note_status(credit_note_status(refund.status))
-        track_refund_status_changed(refund.status)
+        Utils::SegmentTrack.refund_status_changed(refund.status, credit_note.id, organization.id)
 
         result.refund = refund
         result
@@ -50,7 +50,7 @@ module CreditNotes
 
         refund.update!(status:)
         update_credit_note_status(credit_note_status(refund.status))
-        track_refund_status_changed(status)
+        Utils::SegmentTrack.refund_status_changed(refund.status, credit_note.id, organization.id)
 
         if FAILED_STATUSES.include?(status.to_s)
           deliver_error_webhook(message: 'Payment refund failed', code: nil)
@@ -140,18 +140,6 @@ module CreditNotes
         return 'failed' if FAILED_STATUSES.include?(status)
 
         status
-      end
-
-      def track_refund_status_changed(status)
-        SegmentTrackJob.perform_later(
-          membership_id: CurrentContext.membership,
-          event: 'refund_status_changed',
-          properties: {
-            organization_id: organization.id,
-            credit_note_id: credit_note.id,
-            refund_status: status
-          }
-        )
       end
 
       def handle_missing_refund(metadata)
