+++ app/services/credit_notes/update_service.rb
@@ -20,7 +20,7 @@ module CreditNotes
 
       result.credit_note = credit_note
 
-      track_refund_status_changed(credit_note.refund_status)
+      Utils::SegmentTrack.refund_status_changed(credit_note.refund_status, credit_note.id, credit_note.organization.id)
 
       result
     rescue ArgumentError
@@ -30,17 +30,5 @@ module CreditNotes
     private
 
     attr_reader :credit_note, :params
-
-    def track_refund_status_changed(status)
-      SegmentTrackJob.perform_later(
-        membership_id: CurrentContext.membership,
-        event: 'refund_status_changed',
-        properties: {
-          organization_id: credit_note.organization.id,
-          credit_note_id: credit_note.id,
-          refund_status: status
-        }
-      )
-    end
   end
 end
