+++ app/services/invoices/create_one_off_service.rb
@@ -28,7 +28,7 @@ module Invoices
         invoice.finalized!
       end
 
-      Utils::Track.invoice_created(invoice)
+      Utils::SegmentTrack.invoice_created(invoice)
       SendWebhookJob.perform_later('invoice.one_off_created', invoice) if should_deliver_webhook?
       InvoiceMailer.with(invoice:).finalized.deliver_later if should_deliver_email?
       Integrations::Aggregator::Invoices::CreateJob.perform_later(invoice:) if invoice.should_sync_invoice?
