+++ app/services/invoices/paid_credit_service.rb
@@ -26,7 +26,7 @@ module Invoices
         invoice.finalized!
       end
 
-      track_invoice_created(result.invoice)
+      Utils::SegmentTrack.invoice_created(result.invoice)
       SendWebhookJob.perform_later('invoice.paid_credit_added', result.invoice) if should_deliver_webhook?
       InvoiceMailer.with(invoice: result.invoice).finalized.deliver_later if should_deliver_email?
       Integrations::Aggregator::Invoices::CreateJob.perform_later(invoice:) if invoice.should_sync_invoice?
