+++ app/services/invoices/subscription_service.rb
@@ -47,7 +47,7 @@ module Invoices
         Integrations::Aggregator::Invoices::CreateJob.perform_later(invoice:) if invoice.should_sync_invoice?
         Integrations::Aggregator::SalesOrders::CreateJob.perform_later(invoice:) if invoice.should_sync_sales_order?
         Invoices::Payments::CreateService.new(invoice).call
-        Utils::Track.invoice_created(invoice)
+        Utils::SegmentTrack.invoice_created(invoice)
       end
 
       result
