--- app/services/utils/track.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Utils
-  module Track
-    def invoice_created(invoice)
-      SegmentTrackJob.perform_later(
-        membership_id: CurrentContext.membership,
-        event: 'invoice_created',
-        properties: {
-          organization_id: invoice.organization.id,
-          invoice_id: invoice.id,
-          invoice_type: invoice.invoice_type
-        }
-      )
-    end
-  end
-end
